🚚 move: Move GoalForm component with timer functionality and goal management
